+++ R/import_calibrated_data_0.R
@@ -15,6 +15,8 @@ import_calibrated_data_0 <- function(paths) {
     list(data = grep("\\.csv$", l, value = TRUE, ignore.case = TRUE),
          metadata = grep("\\.yml$", l, value = TRUE, ignore.case = TRUE))
 
+  # Exclude Tide Rider CSV files from the data path
+  input_paths$data <- input_paths$data[!grepl(tide_rider_regex, basename(input_paths$data))]
 
   # Check that we found 1 and only 1 of each type of file in the calibration dir
   has_one <- sapply(input_paths, length) == 1
@@ -46,152 +48,23 @@ import_calibrated_data_0 <- function(paths) {
                collapse = "\n\t"), sep = "")
   }
 
-
-  #============================================================================#
-  # Process Metadata                                                        ####
-  #============================================================================#
-
-  # From HOBOware Details.txt files
-  md <- yaml::read_yaml(input_paths$metadata)
-
-  # Define list of expected input md items
-  # TRUE indicates required items,
-  # FALSE is for optional items
-  # Inputs not on this list will be dropped
-  expected_input_md_items <-
-    list(
-      site = FALSE,  # Can be filled in later
-      deployment = FALSE, # Can be filled in later
-      deployment_date = FALSE, # Can be filled in later
-      calibration_start = TRUE, # AKA start of deployment
-      calibration_end = TRUE,  # AKA end of deployment
-      # pct_calibrated  # Calculated later
-      # n_records  # Calculated later
-      # pct_immediate_rejection # later
-      # pct_flagged_for_reviers # later
-      # logging_interval_min # later
-      timezone = TRUE, # use "EST" or "GMT-04:00"
-      # auto_qc_date # later
-      do_calibration = list(n_points = FALSE,  # MX801
-                            pct_saturation = FALSE,
-                            measured_do = FALSE,
-                            temperature = FALSE,
-                            barometric_pressure = FALSE,
-                            start_do_conc = FALSE, # older loggers
-                            start_temperature_c = FALSE,
-                            start_salinity_ppt = FALSE,
-                            start_meter_titration_value_mg_l = FALSE,
-                            start_salinity_correction = FALSE,
-                            end_do_conc = FALSE,
-                            end_temperature_c = FALSE,
-                            end_meter_titration_value_mg_l = FALSE,
-                            start_ratio = FALSE,
-                            end_ratio = FALSE),
-      do_deployment = list( # List for old devices, NA for MX801
-        full_series_name = FALSE,
-        launch_name = FALSE,
-        launch_time = FALSE,
-        calibration_date = FALSE,
-        calibration_gain = FALSE,
-        calibration_offset = FALSE),
-      do_device = list(product = TRUE,
-                       serial_number = TRUE,
-                       version = FALSE),
-
-      cond_calibration = list(
-        date = FALSE,
-        n_points = FALSE,
-        spec_cond_25c = FALSE,
-        measured_cond = FALSE,
-        temperature = FALSE,
-        calibration_points = FALSE,
-        start_cal_cond = FALSE,
-        start_cal_temp = FALSE,
-        end_cal_cond = FALSE,
-        end_cal_temp = FALSE),
-      cond_deployment = list(
-        full_series_name = FALSE,
-        launch_name = FALSE,
-        launch_time = FALSE),
-      cond_device = list(product = TRUE,
-                         serial_number = TRUE,
-                         version = FALSE,
-                         version_number = FALSE,
-                         header_created = FALSE)
-    )
-
-  # Add in optional top level sublists as empty elements if missing
-  optional_top_lists <- c("cond_deployment", "cond_calibration", "do_deployment",
-                          "do_calibration")
-  missing_top_lists <- setdiff(optional_top_lists, names(md))
-  if(length(missing_top_lists) > 0) {
-    miss <- vector(mode = "list", length = length(missing_top_lists))
-    names(miss) <- missing_top_lists
-    md <- c(md, miss)
-  }
-
-  # Check for missing items and eliminate inputs that aren't allowed
-  md <- md[names(md) %in% names(expected_input_md_items)]
-
-  for (i in seq_along(expected_input_md_items)) {
-    n <- names(expected_input_md_items)[i]
-    expected_item <- expected_input_md_items[[n]]
-
-    md_item <- md[[n]]
-    if(is.list(expected_item)) {
-      if(!n %in% names(md)) {
-        stop("Expected top level item: \"", n,
-             "\" missing from YAML file:\n\t", input_paths$metadata, "",
-             sep = "")
-
-      }
-      for(j in seq_along(expected_item)) {
-        inner_name <- names(expected_item)[j]
-           if (expected_item[[j]] && !inner_name %in% names(md_item)) {
-             stop("Required sub-element \"", inner_name, "\" missing from ", n,
-                  "in: ", input_paths$metadata)
-           } # end if required inner item missing
-      } # end loop through sub-list
-
-      # Eliminate extra sublist items
-      if(is.list(md_item)) {
-        md_item <- md_item[names(md_item) %in% names(expected_item)]
-        md[[n]] <- md_item
-      }
-
-    } else {
-      # Not a list, will be logical indicating if it's required
-      if(expected_item && !n %in% names(md)) {
-        stop("Expected top level item:", n,
-             "missing from YAML file:\n\t", input_paths$metadata, "")
-      } # end missing required top level item
-    } # end else
-
-  } # end item loop
-
-  # Create placeholder items which will be filled in later
-  md$pct_calibrated <- NA
-  md$n_records <- NA
-  md$pct_immediate_rejection <- NA
-  md$pct_flagged_for_review <- NA
-  md$logging_interval_min <- NA
-
+  # Process Metadata
+  md <- read_deployment_yaml(input_paths$metadata)
 
   # Read data from file
   d <-  readr::read_csv(file = input_paths$data, show_col_types = FALSE) |>
     clean_csv_import_names()
 
-  d$Date_Time <- format_csv_date_time(d$Date_Time)
+  d$Date_Time <- format_csv_date_time(d$Date_Time, format = "character")
 
   # Record logging interval in minutes
-  t <- d$Date_Time[1:6]
+  t <- d$Date_Time[1:6] |> lubridate::ymd_hms()
   intervals <- (t[2:6]  - t[1:5]) |> as.numeric(units = "mins")
   if(!all(intervals == intervals[1])) {
     stop(input_paths$data, " appears to have a varying interval")
   }
   md$logging_interval_min <- intervals[1]
 
-  d$Date_Time <- as.character(d$Date_Time)
 
   # Other information
   md$site <- paths$site
@@ -201,15 +74,7 @@ import_calibrated_data_0 <- function(paths) {
 
 
 
-  # Check for consistancy in start and end times
-  if(md$calibration_start != as.character(d$Date_Time[1]))
-    stop("Inconsistent start times.")
-
-  if(md$calibration_end != as.character(d$Date_Time[nrow(d)]))
-    stop("Inconsistent end times.")
-
-
-  # Calculate the ratio between calibrartion value from YSI ("TRUTH") to
+  # Calculate the ratio between calibration value from YSI ("TRUTH") to
   #  the uncalibrated logger value
   # Create placeholder items which will be filled in later
   md$pct_calibrated <- NA
@@ -250,9 +115,6 @@ import_calibrated_data_0 <- function(paths) {
   }
 
   d <- d[, final_names]
-
-
-
 
   #----------------------------------------------------------------------------#
   # Read and Process Calibrated DO and Conductivity Tables                  ####